Use 8 bit integers inside Note type

+++ src/note.h
@@ -2,6 +2,11 @@
 #define NOTE_H
 // --------
 
+#include <stddef.h>
+#include <stdint.h>
+
+// --------
+
 #define BASE_FREQUENCY 440.0
 
 // --------
@@ -15,17 +20,17 @@ typedef struct Note Note;
 // --------
 
 struct Note {
-	int pitch;
-	int length;
+	int8_t pitch;
+	int8_t length;
 };
 
 // --------
 
-void Note_init_at(Note* note, int pitch, int length);
+void Note_init_at(Note* note, int8_t pitch, int8_t length);
 
 float Note_get_frequency(const Note* note);
 float Note_get_absolute_length(const Note* note);
-int Note_get_length_in_samples(const Note* note);
+size_t Note_get_length_in_samples(const Note* note);
 
 // --------
 #endif
