+++ src/note_conversion.h
@@ -2,11 +2,13 @@
 #define NOTE_CONVERSION_H
 // --------
 
+#include <stdint.h>
+
 #include "note.h"
 
 // --------
 
-int Convert_musical_to_pitch(char symbol, int half, int* pitch);
+int Convert_musical_to_pitch(char symbol, int8_t half, int8_t* pitch);
 
 // --------
 #endif
